Refactor Atom.swap() and translate swap!
Atom.swap contained a bug that became apparent when you tried to call
swap() with more than 3 arguments; apply() was being called with the IFn
and 3 extra arguments, when apply only takes two arguments, an IFn and a
sequence.

Even still, the use of apply() is deprecated for the use of the splat
*args syntax. Atom.swap() was refactored to take this into account, and
to fix the bug.

swap! was then translated into core.clj.